+++ iShop.Web/Server/APIs/OrdersController.cs
@@ -16,7 +16,7 @@ using Microsoft.Extensions.Logging;
 namespace iShop.Web.Server.APIs
 {
     [Route("/api/[controller]")]
-    public class OrdersController : Microsoft.AspNetCore.Mvc.Controller
+    public class OrdersController : BaseController
     {
         private readonly IMapper _mapper;
         private readonly IUnitOfWork _unitOfWork;
@@ -44,11 +44,16 @@ namespace iShop.Web.Server.APIs
 
         // GET
         [Authorize]
-        [HttpGet("user/{userId}")]
-        public async Task<IActionResult> GetUserOrders(Guid userId)
+        [HttpGet("user/{id}")]
+        public async Task<IActionResult> GetUserOrders(string id)
         {
+            bool isValid = Guid.TryParse(id, out var userId);
+
+            if (!isValid)
+                return InvalidId(id);
+
             if (userId != User.GetUserId())
-                return Unauthorized();
+                return UnAuthorized();
 
             var order = await _unitOfWork.OrderRepository.GetUserOrders(userId);
 
@@ -61,16 +66,20 @@ namespace iShop.Web.Server.APIs
         // GET
         [Authorize]
         [HttpGet("{id}", Name = GetName.Order)]
-        public async Task<IActionResult> Get(Guid id)
+        public async Task<IActionResult> Get(string id)
         {
-            var order = await _unitOfWork.OrderRepository.GetOrder(id);
+            bool isValid = Guid.TryParse(id, out var orderId);
+            if (!isValid)
+                return InvalidId(id);
+            var order = await _unitOfWork.OrderRepository.GetOrder(orderId);
 
             if (order == null)
-                return NotFound(
-                    new ErrorMessage { Code = 404, Message = "item with id " + id + " not existed" }.ToString());
+                return NotFound(ItemName.Order, orderId);
+
             var userId = User.GetUserId();
+
             if (order.UserId != userId)
-                return Unauthorized();
+                return UnAuthorized();
 
             var orderResource = _mapper.Map<Order, OrderResource>(order);
 
@@ -79,6 +88,7 @@ namespace iShop.Web.Server.APIs
 
 
         // POST
+        [Authorize]
         [HttpPost]
         public async Task<IActionResult> Create([FromBody] SavedOrderResource resource)
         {
@@ -91,84 +101,84 @@ namespace iShop.Web.Server.APIs
 
             if (!await _unitOfWork.CompleteAsync())
             {
-                _logger.LogError(LoggingEvents.Fail,
-                    "order " + resource.Id + " with userId " + resource.UserId + " failed to saved");
+                _logger.LogMessage(LoggingEvents.SavedFail, ItemName.Order, order.Id);
 
-                return StatusCode(500,
-                    new ErrorMessage
-                    {
-                        Code = 500,
-                        Message = "order with id " + resource.Id +" by userId " + resource.UserId + " failed to saved"
-                    }
-                        .ToString());
+                return FailedToSave(ItemName.Order, order.Id);
             }
 
             order = await _unitOfWork.OrderRepository.GetOrder(order.Id, false);
             var result = (_mapper.Map<Order, OrderResource>(order));
 
-            _logger.LogInformation(LoggingEvents.Created, "order by userId " + order.UserId + " is created");
+            _logger.LogMessage(LoggingEvents.Created, ItemName.Order, order.Id);
 
             return CreatedAtRoute(GetName.Order,
                 new { id = order.Id }, result);
         }
 
-        // /api/Order/id update a order
+        //PUT
+        [Authorize]
         [HttpPut("{id}")]
-        public async Task<IActionResult> UpdateOrder(Guid id, [FromBody]SavedOrderResource savedOrderResource)
+        public async Task<IActionResult> Update(string id, [FromBody]SavedOrderResource savedOrderResource)
         {
+            bool isValid = Guid.TryParse(id, out var orderId);
+            if (!isValid)
+                return InvalidId(id);
+
             if (!ModelState.IsValid)
                 return BadRequest(ModelState);
 
-            var order = await _unitOfWork.OrderRepository.GetOrder(id);
+            var order = await _unitOfWork.OrderRepository.GetOrder(orderId);
 
             if (order == null)
-                return NotFound(
-                    new ErrorMessage { Code = 404, Message = "item with id " + id + " not existed" }.ToString());
+                return NotFound(ItemName.Order, orderId);
+
             if (order.UserId != User.GetUserId())
-                return Unauthorized();
-            
+                return UnAuthorized();
+
             _mapper.Map<SavedOrderResource, Order>(savedOrderResource, order);
 
             if (!await _unitOfWork.CompleteAsync())
             {
-                _logger.LogError(LoggingEvents.Fail, "item with id " + id + " failed to saved");
-                return StatusCode(500,
-                    new ErrorMessage { Code = 500, Message = "item with id " + id + " failed to saved" }
-                        .ToString());
+                _logger.LogMessage(LoggingEvents.SavedFail, ItemName.Order, order.Id);
+
+                return FailedToSave(ItemName.Order, order.Id);
             }
 
             order = await _unitOfWork.OrderRepository.GetOrder(order.Id);
 
             var result = _mapper.Map<Order, SavedOrderResource>(order);
 
-            _logger.LogInformation(LoggingEvents.Updated, "item with id " + order.Id + " updated");
+            _logger.LogMessage(LoggingEvents.Updated, ItemName.Order, order.Id);
 
             return Ok(result);
         }
 
         // DELETE
+        [Authorize]
         [HttpDelete("{id}")]
-        public async Task<IActionResult> DeleteOrder(Guid id)
+        public async Task<IActionResult> DeleteOrder(string id)
         {
-            var order = await _unitOfWork.OrderRepository.GetOrder(id);
+            bool isValid = Guid.TryParse(id, out var orderId);
+            if (!isValid)
+                return InvalidId(id);
+
+            var order = await _unitOfWork.OrderRepository.GetOrder(orderId);
 
             if (order == null)
-                return NotFound(
-                    new ErrorMessage { Code = 404, Message = "item with id " + id + " not existed" }.ToString());
+                return NotFound(ItemName.Order, orderId);
 
             _unitOfWork.OrderRepository.Remove(order);
 
             if (!await _unitOfWork.CompleteAsync())
             {
-                _logger.LogError(LoggingEvents.Fail, "item with id " + id + " failed to saved");
-                return StatusCode(500,
-                    new ErrorMessage { Code = 500, Message = "item with id " + id + " failed to saved" }
-                        .ToString());
+                _logger.LogMessage(LoggingEvents.SavedFail, ItemName.Order, order.Id);
+
+                return FailedToSave(ItemName.Order, order.Id);
             }
 
-            _logger.LogInformation(LoggingEvents.Deleted, "item with id " + id + " is deleted");
+            _logger.LogMessage(LoggingEvents.Deleted, ItemName.Order, order.Id);
 
-            return Ok(id);
+            return NoContent();
         }
     }
 }
